SoRe 24.04.25 die Buchungstabelle nochmit der Kundentabelle verknüpft und Methode SelctetctKunden hinzugefügt

+++ proj/Kundenverwaltung.xaml.cs
@@ -37,7 +37,7 @@ namespace EasyRentProj
            {
                kunde.vorname = tbVorname.Text;
                kunde.nachname = tbNachname.Text;
-               kunde.nummer = int.Parse(tbNummer.Text);
+               kunde.nummer = (int)long.Parse(tbNummer.Text);
                kunde.adresse = tbAdresse.Text;
                kunde.email = tbEmail.Text;
